+++ packages/api/models/granules.js
@@ -17,7 +17,7 @@ const cmrUtils = require('@cumulus/cmrjs/cmr-utils');
 const log = require('@cumulus/common/log');
 const { getCollectionIdFromMessage } = require('@cumulus/message/Collections');
 const { getMessageExecutionArn } = require('@cumulus/message/Executions');
-const { getMessageCnm, getMessageGranules } = require('@cumulus/message/Granules');
+const { getMessageGranules } = require('@cumulus/message/Granules');
 const { buildURL } = require('@cumulus/common/URLUtils');
 const isNil = require('lodash/isNil');
 const { removeNilProperties } = require('@cumulus/common/util');
@@ -338,7 +338,6 @@ class Granule extends Manager {
       throw new CumulusModelError('meta.collection required to generate a granule record');
     }
 
-    const cnm = getMessageCnm(message);
     const granuleFiles = await FileUtils.buildDatabaseFiles({
       s3,
       providerURL: buildURL({
@@ -380,7 +379,7 @@ class Granule extends Manager {
       timeToArchive: get(granule, 'post_to_cmr_duration', 0) / 1000,
       ...processingTimeInfo,
       ...temporalInfo,
-      cnm,
+      queryFields: get(message, 'meta.granule.queryFields'),
     };
 
     record.published = get(granule, 'published', false);