+++ packages/api/models/schemas.js
@@ -349,22 +349,10 @@ module.exports.granule = {
     dataType: { type: 'string' },
     version: { type: 'string' },
     provider: { type: 'string' },
-    cnm: {
+    queryFields: {
+      description: 'fields for query and metrics purpose',
       type: 'object',
-      properties: {
-        identifier: {
-          type: 'string',
-        },
-        submissionTime: {
-          type: 'string',
-        },
-        receivedTime: {
-          type: 'string',
-        },
-        processCompleteTime: {
-          type: 'string',
-        },
-      },
+      additionalProperties: true,
     },
   },
   required: [